+++ client/keys/add.go
@@ -34,11 +34,6 @@ const (
 	flagNoSort      = "nosort"
 )
 
-const (
-	maxValidAccountValue = int(0x80000000 - 1)
-	maxValidIndexalue    = int(0x80000000 - 1)
-)
-
 func addKeyCommand() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "add <name>",
@@ -201,6 +196,10 @@ func runAddCmd(_ *cobra.Command, args []string) error {
 		if err != nil {
 			return err
 		}
+
+		if !bip39.IsMnemonicValid(mnemonic) {
+			return errors.New("invalid mnemonic")
+		}
 	}
 
 	if len(mnemonic) == 0 {
@@ -214,11 +213,6 @@ func runAddCmd(_ *cobra.Command, args []string) error {
 		if err != nil {
 			return err
 		}
-	}
-
-	if !bip39.IsMnemonicValid(mnemonic) {
-		fmt.Fprintf(os.Stderr, "Error: Mnemonic is not valid.\n")
-		return nil
 	}
 
 	// override bip39 passphrase