+++ x/distribution/tags/tags.go
@@ -5,7 +5,11 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
+// Distribution tx tags
 var (
+	Rewards    = "rewards"
+	Commission = "commission"
+
 	Validator = sdk.TagSrcValidator
 	Delegator = sdk.TagDelegator
 )